# docs/swagger.yaml
definitions:
...
  models.BadRequestResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad request. Missing or invalid required fields.
        type: string
    type: object
...
  models.InternalServerErrorResponse:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal server error occurred.
        type: string
    type: object
...
  models.ServiceResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Operation completed successfully
        type: string
    type: object
  models.StatusUnprocessableEntityResponse:
    properties:
      code:
        example: 422
        type: integer
      message:
        example: Invalid request body format. Please check your JSON syntax and field
          types.
        type: string
    type: object
...
  models.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token upon successful login.
      parameters:
      - description: User login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token and success message
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Invalid credentials or missing fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "422":
          description: Error while parsing the request body
          schema:
            $ref: '#/definitions/models.StatusUnprocessableEntityResponse'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: User Login
      tags:
      - Auth
  /refresh-token:
    post:
      description: Logs out the logged in user and invalidates its existing token.
      produces:
      - application/json
      responses:
        "200":
          description: Refreshed JWT token and success message
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Logout the user
      tags:
      - Auth
...
swagger: "2.0"